valida/requirements.txt
# requirements.txt
streamlit
pytesseract
Pillow
